Give the drain properties a module of their own
`test_subprocess_timeout_properties` had grown past the project's 400-line
ceiling while holding two unrelated concerns: the timeout waiter's deadline
arithmetic and the consumer drain's settlement invariants. Its own docstring
already cited that ceiling as the reason it split from
`test_subprocess_timeout`.

Move the drain properties to `test_subprocess_drain_properties`, along the
seam the section banner already marked. Pure move; both modules are back
under the ceiling, and the drain now has somewhere to gain the capturing
cases it still lacks.
